refactor: rename InformativeScreen to FunnelInformative and updated text- classes in StepHeadline
