+++ packages/storybook/stories/components/onboarding/FunnelInformative.stories.tsx
